+++ CHANGELOG.md
@@ -5,6 +5,14 @@ file is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and
 this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.14.0] - 2023-11-07
+
+### Changed
+
+- Removed REconvergeKindType
+- Moved and renamed crate::ingest::RecordType to crate::RawEventKind
+- Renamed RecordType::Oplog to RawEventKind::OpLog and RecordType::Seculog to RawEventKind::SecuLog
+
 ## [0.13.2] - 2023-11-01
 
 ### Added